+++ app/presenters/export/csv/project_presenter.rb
@@ -255,6 +255,12 @@ class Export::Csv::ProjectPresenter
     I18n.t("export.csv.project.values.single_converter")
   end
 
+  def transfer_type
+    return if @project.type == "Conversion::Project"
+    return I18n.t("export.csv.project.values.form_a_mat") if @project.form_a_mat?
+    I18n.t("export.csv.project.values.single_transfer")
+  end
+
   def esfa_notes
     notes = @project.notes.select do |note|
       note.task_identifier.eql?("update_esfa")